Move ASD predictor/update to own module
Also neatened up the code in a few places.
