+++ stonesoup/updater/kalman.py
@@ -1,5 +1,4 @@
 import warnings
-from collections import OrderedDict
 
 import numpy as np
 import scipy.linalg as la
@@ -8,9 +7,8 @@ from functools import lru_cache
 from ..base import Property
 from .base import Updater
 from ..types.array import CovarianceMatrix
-from ..types.prediction import MeasurementPrediction, ASDGaussianMeasurementPrediction
-from ..types.state import State
-from ..types.update import Update, ASDGaussianStateUpdate
+from ..types.prediction import MeasurementPrediction
+from ..types.update import Update
 from ..models.base import LinearModel
 from ..models.measurement.linear import LinearGaussian
 from ..models.measurement import MeasurementModel
@@ -258,159 +256,6 @@ class KalmanUpdater(Updater):
             hypothesis.prediction,
             posterior_mean, posterior_covariance,
             timestamp=hypothesis.measurement.timestamp, hypothesis=hypothesis)
-
-
-class ASDKalmanUpdater(KalmanUpdater):
-    @lru_cache()
-    def predict_measurement(self, predicted_state, measurement_model=None,
-                            **kwargs):
-        r"""Predict the measurement implied by the predicted state mean
-
-                Parameters
-                ----------
-                predicted_state : :class:`~.ASDState`
-                    The predicted state :math:`\mathbf{x}_{k|k-1}`
-                measurement_model : :class:`~.MeasurementModel`
-                    The measurement model. If omitted, the model in the updater
-                    object is used
-                **kwargs : various
-                    These are passed to :meth:`~.MeasurementModel.function` and
-                    :meth:`~.MeasurementModel.matrix`
-
-                Returns
-                -------
-                : :class:`ASDGaussianMeasurementPrediction`
-                    The measurement prediction, :math:`\mathbf{z}_{k|k-1}`
-
-                """
-        measurement_model = self._check_measurement_model(measurement_model)
-
-        t_index = predicted_state.timestamps.index(
-            predicted_state.act_timestamp)
-        pred_meas = measurement_model.function(
-            State(
-                predicted_state.multi_state_vector[t_index * predicted_state.ndim:
-                                                   (t_index + 1) * predicted_state.ndim]),
-            noise=0, **kwargs)
-
-        hh = self._measurement_matrix(predicted_state=predicted_state,
-                                      measurement_model=measurement_model,
-                                      **kwargs)
-
-        innov_cov = (
-            hh
-            @ predicted_state.multi_covar[t_index * predicted_state.ndim:
-                                          (t_index + 1) * predicted_state.ndim,
-                                          t_index * predicted_state.ndim:
-                                          (t_index + 1) * predicted_state.ndim]
-            @ hh.T + measurement_model.covar())
-
-        meas_cross_cov = predicted_state.multi_covar[
-                         :, t_index * predicted_state.ndim: (t_index + 1) *
-                         predicted_state.ndim] @ hh.T
-
-        return ASDGaussianMeasurementPrediction(
-            multi_state_vector=pred_meas, multi_covar=innov_cov,
-            timestamps=[predicted_state.timestamps[0]],
-            cross_covar=meas_cross_cov)
-
-    def update(self, hypothesis, force_symmetric_covariance=False, **kwargs):
-        r"""The Kalman update method. Given a hypothesised association between
-        a predicted state and an actual measurement,
-        calculate the posterior state. The Measurement Prediction should be
-        calculated by this method. It is overwritten in this method
-
-        Parameters
-        ----------
-        hypothesis : :class:`~.SingleHypothesis`
-            the prediction-measurement association hypothesis. This hypothesis
-            may carry a predicted measurement, or a predicted state. In the
-            latter case a predicted measurement will be calculated.
-        force_symmetric_covariance : :obj:`bool`, optional
-            A flag to force the output covariance matrix to be symmetric by way
-            of a simple geometric combination of the matrix and transpose.
-            Default is `False`
-        **kwargs : various
-            These are passed to :meth:`predict_measurement`
-
-        Returns
-        -------
-        : :class:`~.ASDGaussianStateUpdate`
-            The posterior state Gaussian with mean :math:`\mathbf{x}_{k|k}` and
-            covariance :math:`P_{x|x}`
-
-        """
-
-        # Get the predicted state out of the hypothesis
-        predicted_state = hypothesis.prediction
-
-        # Get the measurement model out of the measurement if it's there.
-        # If not, use the one native to the updater (which might still be
-        # none)
-        measurement_model = hypothesis.measurement.measurement_model
-        measurement_model = self._check_measurement_model(
-            measurement_model)
-
-        # Attach the measurement prediction to the hypothesis
-        hypothesis.measurement_prediction = self.predict_measurement(
-            predicted_state, measurement_model=measurement_model, **kwargs)
-        # Get the predicted measurement mean, innovation covariance and
-        # measurement cross-covariance
-        pred_meas = hypothesis.measurement_prediction.state_vector
-        innov_cov = hypothesis.measurement_prediction.covar
-        m_cross_cov = hypothesis.measurement_prediction.cross_covar
-
-        # Complete the calculation of the posterior
-        # This isn't optimised
-        kalman_gain = m_cross_cov @ np.linalg.inv(innov_cov)
-        posterior_mean = \
-            predicted_state.multi_state_vector \
-            + kalman_gain @ (hypothesis.measurement.state_vector - pred_meas)
-        posterior_covariance = \
-            predicted_state.multi_covar - \
-            kalman_gain @ innov_cov @ kalman_gain.T
-
-        if force_symmetric_covariance:
-            posterior_covariance = \
-                (posterior_covariance + posterior_covariance.T) / 2
-
-        # save the new posterior, if it is no out of sequence measurement
-        try:
-            predicted_state.correlation_matrices[predicted_state.act_timestamp]
-        except KeyError:
-            predicted_state.correlation_matrices[
-                predicted_state.act_timestamp] = {}
-        t_index = predicted_state.timestamps.index(
-            predicted_state.act_timestamp)
-        ndmin = predicted_state.ndim
-
-        # update covariance after calculating
-        predicted_state.correlation_matrices[
-            predicted_state.act_timestamp]['P'] = posterior_covariance[
-                                                  t_index * ndmin:
-                                                  (t_index + 1) * ndmin,
-                                                  t_index * ndmin:
-                                                  (t_index + 1) * ndmin]
-        try:
-            predicted_state.correlation_matrices[
-                predicted_state.act_timestamp]['PFP'] = \
-                predicted_state.correlation_matrices[
-                    predicted_state.act_timestamp]['P'] \
-                @ predicted_state.correlation_matrices[
-                    predicted_state.act_timestamp]['F'].T \
-                @ np.linalg.inv(predicted_state.correlation_matrices[
-                                    predicted_state.act_timestamp]['P_pred'])
-        except KeyError:
-            pass
-        correlation_matrices = OrderedDict(sorted(
-            predicted_state.correlation_matrices.items(), reverse=True))
-
-        return ASDGaussianStateUpdate(multi_state_vector=posterior_mean,
-                                      multi_covar=posterior_covariance,
-                                      hypothesis=hypothesis,
-                                      timestamps=predicted_state.timestamps,
-                                      correlation_matrices=correlation_matrices,
-                                      max_nstep=predicted_state.max_nstep)
 
 
 class ExtendedKalmanUpdater(KalmanUpdater):
